feat: Add additional identity scaffolding items
This commit adds some additional identity scaffolding items that would be useful in a future extension of the project's requirements.

These items were discussed during the user meeting and were initially identified as "won't have" requirements for the initial prototype. However, the scaffolding is now in place in case they need to be utilized in the future.